+++ data/src/environment.rs
@@ -18,11 +18,7 @@ pub fn formatted_version() -> String {
 }
 
 pub fn config_dir() -> PathBuf {
-    portable_dir().unwrap_or_else(|| {
-        dirs_next::config_dir()
-            .expect("expected valid config dir")
-            .join("halloy")
-    })
+    portable_dir().unwrap_or_else(platform_specific_config_dir)
 }
 
 pub fn data_dir() -> PathBuf {
@@ -42,4 +38,30 @@ fn portable_dir() -> Option<PathBuf> {
     dir.join(CONFIG_FILE_NAME)
         .is_file()
         .then(|| dir.to_path_buf())
+}
+
+fn platform_specific_config_dir() -> PathBuf {
+    #[cfg(target_os = "macos")]
+    {
+        xdg_config_dir().unwrap_or_else(|| {
+            dirs_next::config_dir()
+                .expect("expected valid config dir")
+                .join("halloy")
+        })
+    }
+    #[cfg(not(target_os = "macos"))]
+    {
+        dirs_next::config_dir()
+            .expect("expected valid config dir")
+            .join("halloy")
+    }
+}
+
+#[cfg(target_os = "macos")]
+fn xdg_config_dir() -> Option<PathBuf> {
+    let config_dir = xdg::BaseDirectories::with_prefix("halloy")
+        .ok()
+        .and_then(|xdg| xdg.find_config_file(CONFIG_FILE_NAME))?;
+
+    config_dir.parent().map(|p| p.to_path_buf())
 }