+++ src/widget/input.rs
@@ -1,5 +1,7 @@
+use std::collections::HashMap;
+
 use data::user::User;
-use data::{input, Buffer, Command};
+use data::{input, isupport, Buffer, Command};
 use iced::advanced::widget::{self, Operation};
 pub use iced::widget::text_input::{focus, move_cursor_to_end};
 use iced::widget::{component, container, row, text, text_input, Component};
@@ -20,6 +22,7 @@ pub fn input<'a, Message>(
     history: &'a [String],
     users: &'a [User],
     channels: &'a [String],
+    isupport: HashMap<isupport::Kind, isupport::Parameter>,
     buffer_focused: bool,
     disabled: bool,
     on_input: impl Fn(input::Draft) -> Message + 'a,
@@ -35,6 +38,7 @@ where
         input,
         users,
         channels,
+        isupport,
         history,
         buffer_focused,
         disabled,
@@ -66,6 +70,7 @@ pub struct Input<'a, Message> {
     input: &'a str,
     users: &'a [User],
     channels: &'a [String],
+    isupport: HashMap<isupport::Kind, isupport::Parameter>,
     history: &'a [String],
     buffer_focused: bool,
     disabled: bool,
@@ -96,7 +101,9 @@ where
                 // Reset selected history
                 state.selected_history = None;
 
-                state.completion.process(&input, self.users, self.channels);
+                state
+                    .completion
+                    .process(&input, self.users, self.channels, &self.isupport);
 
                 Some((self.on_input)(input::Draft {
                     buffer: self.buffer.clone(),
@@ -162,7 +169,7 @@ where
                         .clone();
                     state
                         .completion
-                        .process(&new_input, self.users, self.channels);
+                        .process(&new_input, self.users, self.channels, &self.isupport);
 
                     return Some((self.on_completion)(input::Draft {
                         buffer: self.buffer.clone(),
@@ -182,9 +189,12 @@ where
                     } else {
                         *index -= 1;
                         let new_input = self.history.get(*index).unwrap().clone();
-                        state
-                            .completion
-                            .process(&new_input, self.users, self.channels);
+                        state.completion.process(
+                            &new_input,
+                            self.users,
+                            self.channels,
+                            &self.isupport,
+                        );
                         new_input
                     };