+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+Changed:
+
+- Simplified onboarding experience for users without a `config.toml` file
+
 # 2024.6 (2024-04-05)
 
 Added: