+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
 - Allow configuration of internal messages in buffer (see [buffer configuration](https://halloy.squidowl.org/configuration/buffer.html#bufferinternal_messages-section))
 
+Changed:
+
+- Simplified onboarding experience for users without a `config.toml` file
+
 # 2024.6 (2024-04-05)
 
 Added: