+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Added:
 
+- Reload configuration file from within the application (<kbd>Ctrl</kbd> + <kbd>r</kbd> (macOS: <kbd>⌘</kbd> + <kbd>r</kbd>))
 - Allow configuration of internal messages in buffer (see [buffer configuration](https://halloy.squidowl.org/configuration/buffer.html#bufferinternal_messages-section))
 - User information added to context menu
 - Support for IRCv3 CAP NEW and CAP DEL subcommands