+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Added:
 - Allow configuration of internal messages in buffer (see [buffer configuration](https://halloy.squidowl.org/configuration/buffer.html#bufferinternal_messages-section))
 - User information added to context menu
 - Support for IRCv3 `CAP NEW` and `CAP DEL` subcommands
-- Enable support for IRCv3 `multi-prefix`, `WHOX`, and `UTF8ONLY`
+- Enable support for IRCv3 `multi-prefix`, `message-tags`, `WHOX`, and `UTF8ONLY`
 - Dynamic commands and tooltips added to command auto-completion via `ISUPPORT` 
 - Added support for `socks5` proxy configuration (see [proxy configuration](https://halloy.squidowl.org/configuration/proxy.html))
 - Added support for `http` proxy configuration (see [proxy configuration](https://halloy.squidowl.org/configuration/proxy.html))